Network.Socket: Add recvBuf
recently I have been looking into a latency-gap between the
C and Haskell implementations of a TCP-echo

One issue that has come up is that recvBufFrom calls getpeername() for each
successful recvfrom().  This is unnecessary in the context of the code that
I am working with as the result is discarded.

My proposed solution to this is to implement recvBuf.

I apologise in advance for my rudimentary Haskell skills.