Fix hClose for SockFD
If we use Network.Socket.close, we get this error:

    <socket: 704>: hClose: user error (close: converted to a Handle, use hClose instead)

Whoops.  SockFD shouldn't call any socket functions that check the MVar.
That, or the user should just use hClose instead of hClose :-)

Also, don't use closeFdWith on Windows.  closeFdWith is currently a no-op
on Windows, but that could change in the future.