Don't use CSize to represent socklen_t
CSize is 64 bits on 64-bit machines, while socklen_t might be 32 bits.
We now use unsigned int (i.e. CUInt) instead, which is what the 64-bit
Linux machine I tested on uses. This is not entirely correct, as C
doesn't guarantee that unsigned long or socklen_t is exactly 32-bits,
but it works in practice.

Closes #81