Test whether recv and send throw exceptions on obviously bogus calls
Due to GHC bug #12010, on x86_64 Windows, in the threaded RTS, the return values
of recv and send aren't checked properly in readRawBufferPtr or
writeRawBufferPtr. Calls that return -1 are interpreted as returning
4,294,967,295 which is then converted back to a 32 bit int equal to -1. This
occurs on all release versions of GHC.

This patch adds tests that run obviously wrong operations and checks that they
throw exceptions.