CI on 64-bit Windows via Appveyor. (#198)
* Commit stack.yaml file.

Useful for hacking on the package reproducibly, and for Windows CI,
which out of convenience is Stack based.

* CI on 64-bit Windows via Appveyor.

Uses Stack as the build and test driver, for convenience. Stack takes
care of installing the right version of GHC and other build tools for
us. Which is rather convenient on Windows given the general lack of
standard package manager.