Officially deprecate the Network module
This module has been kept for backwards compatibility for a long time.
It has been unnoficially deprecated via documentation for a long time
(since 2014), but has never been given a proper deprecation flag. This
flag signals that using this module is an error and should be avoided,
something that can easily be missed by not reading the module summary.