Fix reading SockAddr during `accept` and add tests for Unix sockets
In trying to use Unix domain sockets, we've come across inconsistent failures during `accept`. I tracked it down to the calls to peek the socket address filled by the `accept` syscall. We aren't guaranteed that `sun_path` is approriately null-terminated. This now follows the behavior described in the `unix` man page http://man7.org/linux/man-pages/man7/unix.7.html

This SO that accurately described the behavior I saw: https://stackoverflow.com/questions/17090043/unix-domain-sockets-accept-not-setting-sun-path

I added a new end-to-end test for unix sockets into the test suite. One of the conditions tests for a correctly returned socket address. Without the associated changes, the test fails by getting gobbledygook for the socket address.
I've tested this on OSX as well as centos.