Fix reading SockAddr during `accept` and add tests for Unix sockets
In trying to use Unix domain sockets, we've come across inconsistent failures during `accept`. I tracked it down to the calls to peek the socket address filled by the `accept` syscall. Apparently some unix distros (centos and OSX at the very least) do not fulfill the man spec of setting a null-terminated string in the `sun_path` field for unix sockets.

This SO that accurately described the behavior I saw: https://stackoverflow.com/questions/17090043/unix-domain-sockets-accept-not-setting-sun-path

The changes here make an explicit syscall to get the socket address and appropriately peek it back out.
I added a new end-to-end test for unix sockets into the test suite. One of the conditions tests for a correctly returned socket address. Without the associated changes, the test fails by getting gobbledygook for the socket address.
I've tested this on OSX as well as centos.

I would have prefered to use `getSocketName` from the `Name` module, but the overloaded SocketAddress can't be resolved since it's outside the class instance. Fixing that looked like a larger refactor than fixing this bug. I would prefer to defer refactorings to later diffs as the `network` repo appears to be under heavy development.