Remove EOF errors from receive functions
It is generally accepted that EOF errors in `Network` were a mistake.
Using exceptions for control flow is an anti-pattern in the Haskell
community. While we are in the process of introducing epoch breaking
changes it makes sense to attempt to remove these historic accidents.

There should really be no perceived change for users of
`Network.ByteString(.Lazy).recv`, however `recvFrom` does change
behavior. There are two key differences in `recvFrom`:

1. It is now returning a `(0, address)` on an EOF.
2. It will peek the socket's address and `getPeerName` on an EOF. This
    has some implications for performance, and such may be controversial.

The second point poses a question. Does the return type of
`SocketAddress sa => IO (Int, sa)` make sense for `recvBufFrom`?