Fix regression by manually creating `network.buildinfo` w/o sed
This is needed in order for HsNetworkConfig.h to be installed;
without it the HsNet.h and HsNetDef.h files which do get installed
would become broken for consumers.

We intentionally avoid the use of the AC_CONFIG_FILES macro in order
to reduce the risk of running into with `sed` and Windows' path
separators

See also
https://www.gnu.org/software/autoconf/manual/autoconf.html#File-System-Conventions
for more details about file system related restrictions when using
GNU Autoconf.