Implement socket names, in particular reading a string description of them
Useful for 3rd-party networking applications that might want to pass around
service specifiers without worrying whether these are IP addresses, DNS names,
or UNIX-domain socket paths.

Previously, there was no data type to encapsulate these options together. In
particular, getAddrInfo had to be used to resolve DNS names into a SockAddr
before calling connect/bind, but it could not deal with UNIX domain sockets.
The new function sockNameToAddr takes this role, transparently converting DNS
names and passing through non-DNS-names unaltered, so that it can be used
uniformly without worrying about the specific type of input name/address.

--

I didn't implement a Read instance for SockAddr since readsPrec is supposed to
read in a Haskell expression with a precedence-context and that is incompatible
with how readSockName/readSockAddr is implemented.

In fact the current Show instance for SockAddr already violates the interface,
since it is supposed to output a Haskell expression. Arguably the Show instance
should be derived, and the existing instance should be moved to a top-level
function called something like showSockAddr.

However in theory "readSockAddr 0" should act as pretty much the inverse of
"show" from the current Show instance for SockAddr.