Avoid test race condition
In a few tests the server's socket is not fully "conditioned" when the client
and server threads start.  The server tries to finish socket setup, but this
races against the client thread sending the message, which may be processed by
the kernel with default socket options (observed intermittently in tests on
FreeBSD).

A more complete solution would be to provide setup hooks for the server
setup initialisation, so that the threads don't start until the server
socket is ready.  Ideally the socket options can be set before the
server socket is "bound", ensuring that no messages can come in too
early.  This is not presently warranted, but perhaps some day...