Socket type and family as CInt patterns
The current sum-type model for Socket types and families is not extensible, and
makes it needlessly difficult to perform generic operations on sockets, see
e.g. https://github.com/haskell/network/issues/427

This commit, simplifies the model by replacing the sum-types in question with
newtypes around CInt + patterns for the known constant values.  It also adds
the SO_DOMAIN and SO_PROTOCOL options (when available on the target system).

The "Read" instance for "Family" is for simplicify limited to just the address
families actually supported by the library (unspec, inet, inet6 and unix).
The rest could be added if deemed worth the trouble.