Export StructLinger
StructLinger is useful to set linger option on, with 0 interval using
`setSockOpt`.  This is useful to make 'close' set the RST flag on FIN
packet (reset the connection).  This is not possible with a higher level
`setSocketOption`.

This patch also adds record fields to the `StructLinger` record and
adds haddoc documentation.