Add experimental WinIO support
Another attempt at #364, #509, #602 with the following design choices:

1. On Windows systems, `Socket` becomes an `Either` over the existing
   "Posix" socket implemntation and a Windows equivalent using its
   `SOCKET` type.

2. On Posix systems, or without enabling the WinIO manager with
   `--io-manager=native`, the current behavior is unchanged.

3. *With* the WinIO manager, a greenfield implementation around the
   relevant syscalls is used, using IO Completion Ports
   (AKA "Overlapped IO") through GHC's `withOverlapped`

For the places we need to mux between the two
(`Network.Socket.Syscall`, `Network.Socket.Buffer`, etc.)
the existing implementation is moved into a `Posix.hsc`
and the WinIO one is defined in `WinIO.hsc`.

There's certainly some cleanup and style improvements that could be
done, but it passes all existing tests, and hopefully provides a clean
separation between the existing code (which can stay unchanged) and
direct use of Windows syscalls through `withOverlapped`.